requirements.txt
Babel==1.3
Django==1.6.7
Pillow==2.4.0
South==0.8.4
Unidecode==0.04.16
argparse==1.2.1
dj-static==0.0.6
django-appconf==0.6
django-compressor==1.3
django-extra-views==0.6.5
django-haystack==2.3.1
django-oscar==0.7
django-treebeard==2.0
gunicorn==19.1.1
phonenumbers==6.0.0a
purl==1.0
pytz==2014.7
six==1.8.0
sorl-thumbnail==11.12
static3==0.5.1
wsgiref==0.1.2
